fix(tests): fix broken set() calls from bitcast migration and ASAN workflow path
Three root causes for CI failures on main:

1. Botched bitcast→set() migration (Pattern A): 13 files had inline
   comments inside function call parentheses, e.g.:
     x.set(0, Float32(-100.0  # comment))
   The '#' starts a line comment, consuming the closing '))' and causing
   'expected ) in call argument list' parse errors. Fixed by moving
   comments outside the closing parentheses.

2. Botched bitcast→set() migration (Pattern B/C): test_backward_conv_padding
   had Float32(()) with value on next line; test_rmsprop had garbled
   Float32(= expr); test_optimizers had declarations commented out but
   usage left uncommented; test_core_operations had Int32(Int32()) with
   value on next line.

3. ASAN workflow referenced non-existent path 'tests/shared/core/types'
   (directory does not exist). Fixed to use 'tests/shared/core' with the
   dtype-specific test file pattern.

Affected groups: Core Utilities A, Core Utilities D, Core Gradient,
Core Activations & Types, Shared Infra & Testing, Misc Tests (compile
errors); ASAN Tests (wrong path).

+++ tests/shared/core/test_normalization.mojo
@@ -1255,9 +1255,7 @@ fn test_layer_norm_backward_grad_beta() raises:
     for b in range(3):
         for f in range(4):
             var idx = b * 4 + f
-            grad_output.set(idx, Float32(())
-                Float32(b + 1) * Float32(f + 1) * 0.1
-            )
+            grad_output.set(idx, Float32(b + 1) * Float32(f + 1) * 0.1)
 
     var result15 = layer_norm_backward(grad_output, x, gamma, epsilon=1e-5)
     var grad_beta = result15[2]
